+++ Packages/com.texelsaur.common/Runtime/Scripts/AccessControl.cs
@@ -444,7 +444,7 @@ namespace Texel
             }
         }
 
-        public override bool UsesDebugState => true;
+        public override bool SupportsDebugState => true;
 
         protected override void _UpdateDebugState()
         {