+++ Packages/com.texelsaur.common/Runtime/Scripts/Base/AccessEventBase.cs
@@ -269,5 +269,10 @@ namespace Texel
         {
             get { return accessControl; }
         }
+
+        public virtual bool SupportsStrictSync
+        {
+            get { return false; }
+        }
     }
 }